Version 4049
===== Version 4049 =====
2021/01/14

* Added more functions in lib/math
* Optimized pseudo-random generator
* Bug fixes:
    * Unexpected arithmetic result of float

+++ src/main/java/spl/ast/BinaryOperator.java
@@ -254,7 +254,7 @@ public class BinaryOperator extends BinaryExpr {
                     double leftV = leftTv.floatValue();
                     if (rightTv instanceof SplFloat ||
                             rightTv.isIntLike()) {
-                        double rightV = rightTv.intValue();
+                        double rightV = rightTv.floatValue();
                         result = otherLogical(operator, env, leftV, rightV, getLineFile());
                     } else {
                         return SplInvokes.throwExceptionWithError(
